Makefile update and norminetted some files. Remove too the useless functions like free while we're now using a garbage collector

+++ srcs/parsing/map_parsing.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   map_parsing.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nvaubien <nvaubien@student.42lausanne.c    +#+  +:+       +#+        */
+/*   By: rrouille <rrouille@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/24 16:00:32 by nvaubien          #+#    #+#             */
-/*   Updated: 2023/12/16 00:31:11 by nvaubien         ###   ########.fr       */
+/*   Updated: 2023/12/16 12:42:03 by rrouille         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,7 +70,7 @@ int	contains_only_valid(char *line)
 
 int	contains_1_start_end(char *line)
 {
-	int	i;
+	unsigned long	i;
 
 	i = 0;
 	while (line[i])
